REF SECTION UPDATE:
Modified the Ref page by changing the container tag to a div.
Created 3 div with a className: marquee that contains the Attribution contents.
Created each Attribution content as a list item within and unordered list marquee-content.
Added a useEffect to help implement the duplication effect of the marquee content.
This is donme by getting the specified amount of marquee-elements to be dispalyed, appending them into an array, looping through & duplicating them.
This is done to ensure the marquee type effect is done on each list item without displaying any blank spaced a they slide through the screen.
Added the appropriate styling and animation effect using CSS animation.